backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

fastapi==0.118.0
uvicorn==0.37.0
requests==2.32.5
python-dotenv==1.1.1
openai==2.1.0
loguru==0.7.3
pydantic-settings==2.10.1
transformers==4.57.0
torch==2.9.0
kernels==0.10.2
accelerate==1.10.1
hf_xet==1.1.10
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10
pydantic==2.11.7
sqlmodel==0.0.27
prisma==0.15.0
passlib==1.7.4
pydantic[email]==2.11.7
bcrypt==4.0.1
python-jose==3.5.0
alembic==1.17.0
langchain-openai==0.3.35
python-multipart==0.0.20